updates from testing in dec 2020
The .env and manifest were still using the username/password
convention. This PR changes those files to api key and URL,
which are already used in the code base.

Additionally, one minor change to make the message a bit more
useful when an error is returned. For instance, now it'll
show the the model is still being trained, rather than a
not useful, even if comical, frowny face.